+++ include/ambition/7kaaInterface/control.hh
@@ -46,6 +46,9 @@ void displayNews(
 void finishReplay(
 );
 
+void initialiseErrorReporting(
+);
+
 /**
  * Migrate files from the everything-in-one config directory to the separate
  * ones.
@@ -114,6 +117,9 @@ bool shouldSelectSortiedUnits(
  */
 void unlockBuffer(
   VgaBuf& buffer
+);
+
+void useGraphicalErrorReporting(
 );
 
 } // namespace _7kaaAmbitionInterface::Control