+++ include/ambition/Ambition_control.hh
@@ -1,7 +1,7 @@
 /*
  * Seven Kingdoms: Ambition
  *
- * Copyright 2025 Tim Sviridov
+ * Copyright 2025–2026 Tim Sviridov
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -62,6 +62,9 @@ void openFeedback(
 );
 
 void startMusic(
+);
+
+void useGraphicalErrorReporting(
 );
 
 } // namespace Ambition::Control