+++ src/ambition/version_details.cc
@@ -22,6 +22,12 @@ std::vector<VersionUpdates> getVersionDetails(
     .identifier = IDENTIFIER_PREVIEW,
     .sections = {
       {
+        .heading = HEADING_ADDED,
+        .items = {
+          _("Show a message to the user when Seven Kingdoms: Ambition crashes instead of just disappearing, to aid in reporting bugs.  If you do receive a crash message, please take a screenshot and report it."),
+        },
+      },
+      {
         .heading = HEADING_FIXED,
         .items = {
           _("Fixed caravans with no stops not being counted as idle, and so never being retired or re-used."),
@@ -40,6 +46,7 @@ std::vector<VersionUpdates> getVersionDetails(
           _("Caravans now show as idle (with an asterisk) on the Trade Report when they are considered idle by the automatic caravan replacement and cloning system."),
           _("Draw ship washes (the waves around, and the wake behind, ships) at a constant FPS, independent of game speed."),
           _("Do not show ship washes for ships that are stopped and not moving."),
+          _("Made development slightly easier."),
         },
       },
     },