+++ CHANGELOG.md
@@ -10,6 +10,12 @@ speaking, considered to be the player's experience.
 
 ## [Unreleased]
 
+### Added
+
+- Show a message to the user when Seven Kingdoms: Ambition crashes instead of
+  just disappearing, to aid in reporting bugs.  If you do receive a crash
+  message, please take a screenshot and report it.
+
 ### Fixed
 
 - Fixed caravans with no stops not being counted as idle, and so never being
@@ -36,6 +42,7 @@ speaking, considered to be the player's experience.
 - Draw ship washes (the waves around, and the wake behind, ships) at a constant
   FPS, independent of game speed.
 - Do not show ship washes for ships that are stopped and not moving.
+- Made development slightly easier.
 
 
 ## [3.2.12] — 2026-09-01